+++ web_generator/bin/update_bindings.dart
@@ -64,7 +64,7 @@ $_usage''');
   }
 
   // Determine the set of previously generated files.
-  final domDir = Directory(p.join('..', 'web', 'lib', 'src', 'dom'));
+  final domDir = Directory(p.join('..', '..', 'web', 'lib', 'src', 'dom'));
   final existingFiles =
       domDir.listSync(recursive: true).whereType<File>().where((file) {
     if (!file.path.endsWith('.dart')) return false;
@@ -79,7 +79,7 @@ $_usage''');
   // Run app with `node`.
   await _runProc(
     'node',
-    ['main.mjs', '../../web/lib/src'],
+    ['main.mjs', '../../../web/lib/src'],
     workingDirectory: _bindingsGeneratorPath,
   );
 
@@ -93,7 +93,7 @@ $_usage''');
 
   // Update readme.
   final readmeFile = File(
-    p.normalize(p.join(_bindingsGeneratorPath, '..', 'README.md')),
+    p.normalize(p.join(_bindingsGeneratorPath, '..', '..', 'README.md')),
   );
 
   final sourceContent = readmeFile.readAsStringSync();
@@ -133,13 +133,13 @@ String _packageLockVersion(String package) {
   return webRefIdl['version'] as String;
 }
 
-const _bindingsGeneratorPath = 'generator';
+const _bindingsGeneratorPath = '../lib/src';
 
 const _webRefCss = '@webref/css';
 const _webRefElements = '@webref/elements';
 const _webRefIdl = '@webref/idl';
 
-const _thisScript = 'update_bindings.dart';
+const _thisScript = 'bin/update_bindings.dart';
 
 const _startComment =
     '<!-- START updated by $_thisScript. Do not modify by hand -->';
@@ -169,7 +169,7 @@ Future<void> _runProc(
 Future<void> _generateJsTypeSupertypes() async {
   // Use a file that uses `dart:js_interop` for analysis.
   final contextCollection = AnalysisContextCollection(includedPaths: [
-    p.fromUri(Platform.script.resolve('../web/lib/src/dom.dart'))
+    p.fromUri(Platform.script.resolve('../../web/lib/src/dom.dart'))
   ]);
   final dartJsInterop = (await contextCollection.contexts.single.currentSession
           .getLibraryByUri('dart:js_interop') as LibraryElementResult)